requirements.txt
Django==5.2.5
djangorestframework
gunicorn
dj-database-url
mysqlclient
whitenoise
python-dotenv
pdfplumber
pdf2image
pytesseract
Pillow
google-generativeai